Squashed commit of the following:

commit ec230de1016fe57e421ea2899ccdd2a874ddca65
Author: Shravan Rangaraju <shrarang@cisco.com>
Date:   Wed Oct 6 07:18:46 2021 -0400

    appid: fix update_allocations signature in unit test

+++ src/network_inspectors/appid/test/appid_efp_process_event_handler_test.cc
@@ -54,7 +54,7 @@ Packet::Packet(bool) { }
 Packet::~Packet() = default;
 }
 
-void memory::MemoryCap::update_allocations(unsigned long) { }
+void memory::MemoryCap::update_allocations(size_t) { }
 void memory::MemoryCap::update_deallocations(size_t) { }
 
 void ApplicationDescriptor::set_id(const Packet&, AppIdSession&, AppidSessionDirection,
